make it compile with stricter warnings

+++ include/string.h
@@ -2,9 +2,9 @@
 #include <stddef.h>
 #include <stdint.h>
 // set `n` bytes of `s` to `c`
-extern void   *memset(void *s, uint_fast64_t c, size_t n);
+extern void   *memset(void *s, uint8_t c, size_t n);
 extern void   *memcpy(void *destination, const void *source, size_t num);
-extern int     memcmp(const void *s1, const void *s2, int_fast64_t len);
+extern int     memcmp(const void *s1, const void *s2, size_t len);
 extern void   *memmove(void *dest, const void *src, size_t n);
 extern int     strcmp(const char *str1, const char *str2);
 extern char   *strcat(char *restrict destination, const char *restrict source);
