+++ libc/string/memcmp.c
@@ -1,5 +1,6 @@
 #include <stdint.h>
-int memcmp(const void *s1, const void *s2, int_fast64_t len) {
+#include <stddef.h>
+int memcmp(const void *s1, const void *s2, size_t len) {
 	const uint8_t *p = s1;
 	const uint8_t *q = s2;
 	int charCompareStatus = 0;
